Introduce abstract Transaction interface to support off-chain operations
Extracted a generic `Transaction` interface and introduced the
`OnchainTransaction` class.

This abstraction prepares the codebase for supporting off-chain
transactions, such as operations imported from CEX platforms like
Binance. A swap, for instance, will be represented as two transactions:
one outgoing and one incoming, regardless of the origin (on-chain or
off-chain).